Support handlebars 2,3,4.
I have refrained from specifying handlebars as a peerDependency for a
few reasons:

1) in npm@2, peerDependencies cause side effects on the dependency tree,
meaning dependency hell is a problem: it forces all packages, even with
their devDependencies, to use the same version of packages. This is not
realistic.

2) in npm@3, peerDependencies simply print a warning when unfulfilled.
This is superb, but I'm betting that many users have not yet
upgraded to npm@3 yet.